Setup the relation manager for prospect tasks
